Move types from VendorItemCard to types file
